+++ src/grids/content-types-grid/content-types-list-modal/content-types-list-modal.module.ts
@@ -3,13 +3,13 @@ import { NgModule, ModuleWithProviders } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { ContentTypesListModalComponent } from './content-types-list-modal.component';
 import { ModalModule } from 'ngx-bootstrap';
-import { ModalFooterButtonsModule } from './../../../controls/modal-footer-buttons/modal-footer-buttons.module';
+import { FooterButtonsModule } from './../../../controls/footer-buttons/footer-buttons.module';
 import { TranslateModule } from '@ngx-translate/core';
 
 @NgModule({
   imports: [
     CommonModule, ModalModule.forRoot(), TranslateModule.forChild(),
-    ModalFooterButtonsModule.forRoot(), ContentTypesGridModule.forRoot()
+    FooterButtonsModule.forRoot(), ContentTypesGridModule.forRoot()
   ],
   declarations: [
     ContentTypesListModalComponent