+++ src/grids/permissions-grid/permissions-list-modal/permissions-list-modal.module.ts
@@ -4,12 +4,12 @@ import { PermissionsListModalComponent } from './permissions-list-modal.componen
 import { ModalModule } from 'ngx-bootstrap';
 import { PermissionsGridModule } from './../permissions-grid.module';
 import { TranslateModule } from '@ngx-translate/core';
-import { ModalFooterButtonsModule } from '../../../controls/modal-footer-buttons/modal-footer-buttons.module';
+import { FooterButtonsModule } from '../../../controls/footer-buttons/footer-buttons.module';
 
 @NgModule({
   imports: [
     CommonModule, ModalModule.forRoot(), TranslateModule.forChild(),
-    ModalFooterButtonsModule.forRoot(), PermissionsGridModule.forRoot(), PermissionsGridModule.forRoot()
+    FooterButtonsModule.forRoot(), PermissionsGridModule.forRoot(), PermissionsGridModule.forRoot()
   ],
   declarations: [
     PermissionsListModalComponent