+++ src/grids/users-grid/account-profile-form/account-profile-form.module.ts
@@ -4,13 +4,13 @@ import { AccountProfileFormComponent } from './account-profile-form.component';
 import { UserGroupsGridModule } from './../../user-groups-grid/user-groups-grid.module';
 import { TextInputModule } from './../../../controls/text-input/text-input.module';
 import { CheckboxesInputModule } from './../../../controls/checkboxes-input/checkboxes-input.module';
-import { ModalFooterButtonsModule } from './../../../controls/modal-footer-buttons/modal-footer-buttons.module';
+import { FooterButtonsModule } from './../../../controls/footer-buttons/footer-buttons.module';
 import { FormsModule } from '@angular/forms';
 
 @NgModule({
   imports: [
     CommonModule, FormsModule, TextInputModule.forRoot(),
-    CheckboxesInputModule.forRoot(), ModalFooterButtonsModule.forRoot(),
+    CheckboxesInputModule.forRoot(), FooterButtonsModule.forRoot(),
     UserGroupsGridModule.forRoot()
   ],
   declarations: [