+++ src/grids/users-grid/user-modal/user-modal.module.ts
@@ -2,7 +2,7 @@ import { NgModule, ModuleWithProviders } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { UserModalComponent } from './user-modal.component';
 import { ModalModule } from 'ngx-bootstrap';
-import { ModalFooterButtonsModule } from './../../../controls/modal-footer-buttons/modal-footer-buttons.module';
+import { FooterButtonsModule } from './../../../controls/footer-buttons/footer-buttons.module';
 import { UserGroupsGridModule } from './../../user-groups-grid/user-groups-grid.module';
 import { TextInputModule } from './../../../controls/text-input/text-input.module';
 import { CheckboxesInputModule } from './../../../controls/checkboxes-input/checkboxes-input.module';
@@ -12,7 +12,7 @@ import { FormsModule } from '@angular/forms';
 @NgModule({
   imports: [
     CommonModule, FormsModule, ModalModule.forRoot(), TranslateModule.forChild(),
-    ModalFooterButtonsModule.forRoot(), UserGroupsGridModule.forRoot(),
+    FooterButtonsModule.forRoot(), UserGroupsGridModule.forRoot(),
     TextInputModule.forRoot(), CheckboxesInputModule.forRoot()
   ],
   declarations: [