+++ src/grids/users-grid/users-list-modal/users-list-modal.module.ts
@@ -1,7 +1,7 @@
 import { NgModule, ModuleWithProviders } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { UsersListModalComponent } from './users-list-modal.component';
-import { ModalFooterButtonsModule } from './../../../controls/modal-footer-buttons/modal-footer-buttons.module';
+import { FooterButtonsModule } from './../../../controls/footer-buttons/footer-buttons.module';
 import { UsersGridModule } from './../users-grid.module';
 import { ModalModule } from 'ngx-bootstrap';
 import { TranslateModule } from '@ngx-translate/core';
@@ -9,7 +9,7 @@ import { TranslateModule } from '@ngx-translate/core';
 @NgModule({
   imports: [
     CommonModule, ModalModule.forRoot(), TranslateModule.forChild(),
-    ModalFooterButtonsModule.forRoot(), UsersGridModule.forRoot()
+    FooterButtonsModule.forRoot(), UsersGridModule.forRoot()
   ],
   declarations: [
     UsersListModalComponent